Port of Cosmins SML, using happy/alex parser generator.
There are a few differences; for example I do not permit
single-element tuples in the syntax.  This was likely a bug anyway.
The parser still has a ton of shift-reduce conflicts due to ambiguous
syntax.  We'll have to take a closer look.

I will write a type checker before I write an interpreter, probably
tomorrow.  This is likely to involve small changes to the abstract
syntax.

Finally, the parser and lexer do not perform particularly elegant
error reporting.  This should probably be corrected at some point.