Finished typechecker.
Not everything is tested, but it can handle the test programs.

Operator-lambdas (op+ etc) are a hack, due to the need for
polymorphism.

The position information is stored in a really ugly way.  The abstract
syntax should probably be changed again to wrap its components in some
information-carrying functor, like in GHC.