Add file output support to the --verbose flag.
--verbose (or -V) now takes an option specifying a file on which to
put the final, erroneous program.  This is useful when debugging the
compiler, as you can run e.g.

  % l0c -<flags> -Vbad.l0 prog.l0
  Running foo.
  Running bar.
  ...
  Error during pass 'foobar':
  ...

after which the input program to the failing pass will be found in
bad.l0.  Previously, you had to pipe standard error to a file, and
manually remove the first few lines of compiler output.  It's much
easier now.