Change syntax of tuples and arrays.
Arrays are now written as [1,2,3], tuples as {x,y,z}, as in Erlang.
The change to tuples also affects pattern and type syntax.
Additionally, tuples with one and zero elements are now supported.
shape() now always produces a tuple (a one-tuple in the case of a
one-dimensional array).

Next up: make SOACs always return tuples as well.