Add proper Assert primitive to the language.
Asserts take a boolean argument, which, if false, causes execution to
halt.  They return certificates (currently boolean values), which
other expressions may be predicated upon.

It's a bit complex.  Ask me if you're in doubt.

There's still a bunch of vestigial assertZip stuff around, I'll work
on removing that.

I also need to teach the copy propagator about asserts, because
there's currently a lot of redundancy that it could remove.