Change assertions to return type 'cert' instead of 'bool'.
The cert type has only a single value, Checked.  The built-in function
'conjoin' can be used to convert several cert values into a single
one: conjoin(c1,c2,c3).  Any number of arguments is accepted.

The reason for this change is that it never made much sense for
certificates to be represented as booleans, given that they could
never be false.  With their own type, we can also remove them later on
in the code generation, and thus have no run-time representation.