Change SubExps constructor to SubExp.
The former was just a potential source of mischief, with the dubious
benefit of being slightly more expressive.

I also took out the "alternatives expression" support in the hoister.
It had been dead code for a while, since we originally used it to
remove fusion nuisances, but that is now handled in a different way.