split: allow more than one split point
In the external language you can now do
@split( (1,1,3), [ 1, 2, 3, 4 ]) = {[1], [], [2, 3], [4]}@

Here indexes are used to split the array, and you will get an
array with the leftover elements.

The index expressions are called "splitexps" in the code.

---

In the internal language, we use sizes to build "multi-take"
@a = [1,2,3,4]@
@split( (1,0,2) , a ) = {[1], [], [2,3]}@

Note that you do _not_ get an extra array with leftover elements.

The size expressions are called "sizeexps" in the code.