Remove the ! notation for shape declarations.
A shape declaration in a parameter is now always binding, and a shape
declaration in a return type must now always refer to a name in scope.

This is strictly less powerful than before - it is not possible to
indicate, for example, that a function returns a square array of
arbitrary size - the return shape declarations must consist entirely
of parameter names (or not be given).  This can be improved later.