Options for selecting the OpenCL device to run on.
The C/OpenCl code generator will now generate programs that accept
`--platform` (`-p`) and `--device` (`-d`) options.  These options each
take an argument that specifies a subset of the name of the platform
or device that Futhark will attempt to execute on.  For example, one
could pass `--device Hawaii` to use the AMD GPU on napoleon.  Use the
`clinfo` tool to figure out the sometimes obtuse names.

Closes #87.