Begin work on compiling reduction streams more efficiently.
The trick here is to interchange the loops when compiling reductions
containing maps - we already do this in fusion, now we will also do it
when extracting kernels.

This required a new kernel construct - ChunkedMapKernel.  This is not
really a map, but more like the first part of a two-part reduction
kernel.