Introduce an `include` header for Futhark programs.
I need this.  My implementation changes as little as possible.  In particular,
all internal data structures are still the same.  The downside is that the
parser now has a slightly larger job to do, and part of that job doesn't have
anything to do with parsing.  I'll salute anyone who wishes to do this properly
at some point.