Modify the Write SOAC and kernel to use a Lambda.
This works with the interpreter and the C code generator (through the first
order transform), but only works to some extent in the generated OpenCL code.
For values of non-primitive types, the generated code throws a "failed with
error code -51 (Invalid argument size)" error.  I think this might be because of
my shady ExplicitAllocations.hs code, which I'm not really sure what is doing.