WIP - needs bugfixing
filterIndexes used to just find all index operations. Now it finds which
kernel variables we're interested in as well and stores it right in the
Access values themselves. This is useful, because it minimises the work
the algorithm has to apply to the data after the summary has been
created.

Furthermore, the summary itself should in essence be a list of accesses
that each sum up all relevant information to themselves. The algorithm
for determining coalescing should then take this list of summaries and
choose a strategy from that.
I'll argue that distributing the information for the summary over
individual accesses is a good way of capturing the relevant information
that our task needs.