Do not check futhark-mode.el in Travis.
The Ubuntu 12.04 version of Travis does not contain a new enough Emacs
in its package archive for this to work.  We need at least Emacs 24,
but can at most get Emacs 23.  We *could* make emacs from source, but
it seems very overkill.

I should have done this in a branch...