Extend the experimental memory block merging optimisation.
Now set the environment variable `MEMORY_BLOCK_MERGING` to `1` to
enable the optimisation in all tools.

Also:

  + Add more tests, and move the existing ones a little around.
  + Make the debug printing more readable.

Running `MEMORY_BLOCK_MERGING=1 futhark-test tests/memory-block-merging`
shows that not all the 27 tests in the directory pass with the optimisation
enabled.  Some of the bugs seem to have easy fixes, but some of them maybe
don't.  (They all pass without the optimisation enabled.)

This is very much work in progress.