Enable memory block merging in sequential code generation.
All tests and benchmarks validate with futhark-c and the current state
of the optimisation, so it is time.

Also use the new 'structure cpu' attribute on several memory block
merging tests to describe the expected number of allocations in an
optimised program.

Since the futhark tools should not spew out text by default, you now
have to opt-in to get the debug prints in the memory block merging
modules.  Export the environment variable FUTHARK_DEBUG to see the
debug prints. @coancea

With futhark --gpu, you still have to set MEMORY_BLOCK_MERGING=1 to
enable the experimental optimisation.