Make memory block merging almost work with the current compiler
Extend futhark-bench.hs to also save stderr in a JSON file.  This is useful when
used with '--pass-option=-debugging'.

Still fails on some programs.  It is still disabled by default in the the
Futhark.Passes module.

Still has a lot of debbuging code, but should be easy to clean up once deemed
superfluous.

Fix a bug where the order of creation of arrays used in a concat could
artificially restrict coalescing (see
memory-block-merging/tests/coalescing/concat/pos3.fut).