Make memory block merging work with the current compiler
Extend futhark-bench.hs to also save stderr in a JSON file.  This is useful when
used with '--pass-option=-debugging'.

Works with all benchmarks and test programs (that also work without it).  There
are still a few cases where the resulting code is poorer in terms of speed or
peak memory usage.  The optimisations are still disabled by default in the the
Futhark.Passes module.  The tests are not moved into the general tests directory
yet.

Remove all memory block merging debugging code.

Move all memory block tests to the general tests directory.

Rename the *Hoisting modules.

Be a bit more conservative in the last use analysis.

Disable memory reuse on constant-sized allocations.  Makes the heston32
benchmark not segfault, but I'm not entirely sure why.

Create a function offsetIndexDWIM that handles edge cases.

Fix misc. edge cases in MemoryUpdater.

Fix the PrimExp Eq instance wrt. to literals.