Free some allocations earlier.
This introduces the use of basic liveness information (based on freeIn
on the generated ImpCode) to potentially free allocations that go dead
within the same sequence of statements.

Due to the way the allocator works for generated C code, it will not
have much actual benefit there (but the stats will still improve).