Use the host-computed workgroup size even within the kernel.
This has the strong advantage that frequently, this size is a
kernel-compile-time constant, which permits the kernel compiler to
perform various simplifications.

The reason we used to query the actual workgroup size directly was in
the naive hope that the kernel compiler would be able to use this
information for something useful.  Time (and benchmark results) will
tell if this ever did anything.