Implement a new interpreter for Futhark.
This interpreter operates directly on the source language, and
implements it faithfully without regard to any specific implementation
technique.  This does cause some problems - I am quite confident there
are exotic problems related to interactions between size annotations
and empty arrays, that are handled differently by the compiler and
interpreter.

This interpreter is *very* slow.  It is not suitable for running
programs on all but the smallest data sets.  It may be possible to
speed it up in the future.  On the upside, it supports a very nice
interactive environment that supports the full language, including
modules.

It completely replaces (and removes) the old core language
interpreter.

Closes #537.