Add SegRed kernel for #666.
This represents all reductions, segmented or not, with a single
specialised construct (SegRed).  This is turned into the usual
multiversioned mess in the code generator.  The main advantage of this
representation is simplicity of both the IR and the generated code.

Performance is the same for our application benchmarks.  For
microbenchmarks, it varies a bit, but can be slower for small
workloads (the ones that run in a few dozen microseconds anyway).

I believe a similar ScanRed kernel could be implemented, and would go
quite a way towards addressing #666.