Enable the LMAD representation.
An index function is now represented as a list of LMADs.

This is a drop-in replacement for the old index function
representation.  There are not yet any optimisations that rely
directly on the LMAD representation.

Add unit tests and a very simple algebraic index function
representation to compare against.

The new LMAD module also fixes a few bugs in the previous Lmad module,
and adds a few more heuristics.

Adjust memory expansion to recompute LetAttr's where possible.  This
reduces the number of rebase operations needed, which makes it easier
to unify index functions in the type checker.

Remove all of the old memory block merging optimisations.  These
optimisations do not readily work with the LMAD representation (we
could tweak them to kind of keep working for a while longer, but I
don't really want to).