Fix #728.
This is a poor fix, and is likely to break down in more complex
circumstances (my guess: tuples that mix unique and non-unique
arrays).

The core problem is that our type representation does not let us
represent the case where the return value may alias certain of the
parameters, but not others.