Disable the inlining optimisation in the defunctionaliser.
The way we did it, we were obscuring alias information that we use to
infer uniqueness attributes for lifted functions (which is probably a
broken idea, but let's catch that fish another day).

It doesn't look like it had much impact on compiler performance
anyway.  I'd rather improve the inliner and (especially) core type
checker instead.