Generate fewer replicate kernels.
The idea is to detect the overwhelmingly common case (filling an
entire row-major array with a primitive value) and generate a single
function with a single kernel for each type that is used.  This is
exactly the same idea that we do for transpositions (except that the
transpose kernels are of course much more complicated).