Index function generalization/antiunification across if-then-else
The purpose of this commit is to decrease copies in if-then-else branches.

Before, if two branches of an if-then-else returned arrays with different index
functions (LMADs), the compiler would normalize the arrays in one or both of the
branches, which could result in unnecessary copying.

With this commit, the compiler will try to generalize the index functions and
return the existentials in addition to the array itself. Consider the following
expression:

```futhark
  if (arr[0] < 0.0)
  then arr[2+n/2:n]
  else arr[2+n/4:n]
```

Before, the results would be copied into fresh arrays in both branches, but
really, we could return the same array as before in addition to the computed
offsets to the index function. This way, we can avoid some unnecessary copies.

Along the way, we've had to relax the type checker for the ExplicitMemory
representation a bit. To illustrate why, consider the expression above. In some
cases the simplifier will hoist the computation of the offset to outside the
generated branches, meaning that the type checker cannot know that the offset
computed and the offset of the returned array match. We therefore introduce the
function `closeEnough` which we hope is good enough to catch future errors in
the code generation, but which is relaxed enough to let this particular case
through.

We also need to take care when generalization index functions where the same
value is used in different places: For instance, given an index function with
base and shape `x`, and another index function with base and shape `y`, the
generalization would produce two existentials `0 => (x, y)` and `1 => (x, y)`,
instead of just one. This commit first checks if a given substitution already
exists, in which case it will try to reuse it. Except for the size parameter or
shape of the MemArray, which we have separated from the shape reported in the
index function LMAD.

It is also important that the memory existential in ReturnsNewBlock are numbered
_after_ the other existentials, even when multiple blocks are being returned.
For instance, `([?0]@?1->..., [?2]@?3->...)` is not a valid return type, it
should instead be `([?0]@?2->..., [?1]@?3->...)`.

We also ran into a case in ImpGen where the assumption that all
allocations (outside of the return of lambda functions) are explicit in the IR
wasn't being respected.

The changes pass all unit tests and seem to have no regression in benchmark
performance. There doesn't seem to be any speedup either, but hopefully this
change will pave the way for future changes that have more impact on efficiency
of the generated code.