Fix existential `if`s
The commit ac2701e217310845102cb96408b1b875cd3f8f5b introduced an optimization
that allowed Futhark to reduce the amount of copies across if-branches by using
index function existentials. The idea was good, but the implementation
introduced bugs that reduced in for instance the sorts tests to fail:

https://travis-ci.org/diku-dk/sorts/builds/655258023

First of all, the previous generalization would erroneously generalize the same
pair of values to different existentials if they were present in different parts
of the index function: For instance, given an index function with base and shape
`x`, and another index function with base and shape `y`, the generalization
would produce two existentials `0 => (x, y)` and `1 => (x, y)`, instead of just
one. This commit first checks if a given substitution already exists, in which
case it will try to reuse it.

We also handled the existentials as sets, but the order in which they are
returned is important in order to match the pattern generated. Therefore, we
have amended the generalization functions to return substitutions as an ordered
set.

It is also important that the memory existential in ReturnsNewBlock are numbered
_after_ the other existentials, even when multiple blocks are being returned.
For instance, `([?0]@?1->..., [?2]@?3->...)` is not a valid return type, it
should instead be `([?0]@?2->..., [?1]@?3->...)`.