Implement binary searching in Autotuner
The purpose of this PR is to improve the autotuner by taking into account
programs with changing levels of parallelism. Previous to this commit, if a
program had different levels of parallelism, called `ePar`, the autotuner would
choose the first one it encountered. In the LUD benchmark, that would always be
the highest `ePar`, since the level of parallelism is decreasing. But how do we
know that this is the best ePar to compare against?

Assuming that program performance is monotonic in the threshold value chosen, we
can instead try to find the best `ePar` by binary search over all the possible
values. Of course, autotuning will take longer if there are more `ePar`s, but
this is not the case for most of our benchmarks, and for those that it is, we
hope that we can get a better `ePar` as a result.

The runtime for the autotuner should only increase by `log(n)`, where `n` is the
number of `ePar`s in a given program. Tuning LUD goes from 163s to 300s on
gpu04.

For LUD, I'm getting a good speedup on bigger input data:

```
lud.fut
  data/512.in:                                                          2.16x
  data/64.in:                                                           0.99x
  data/256.in:                                                          1.35x
  data/16by16.in:                                                       1.00x
  data/2048.in:                                                         1.47x
```