Always pick the best runtime when tuning
Futhark is deterministic, so we should be able to use the best runtime as a
benchmark to determine the tuning parameters. Doing so means that we ignore all
runs where something was disturbing the benchmark run, such as other processes
using the GPU. We don't want such cases to disrupt our tuning. Hopefully this
change will make tuning results more consistent.