Do not reuse the original pattern in these simplification rules.
It can be dangerous to reuse the original pattern, because the pattern
element annotations (most importantly the aliases) may no longer be
correct.  This reflects a design flaw in our system.  Maybe we should
rip aliases out of those attributes entirely, or else the simplifier
should be careful to recompute them when it needs them.  They contain
only local information (with the exception of 'if'... sigh), so it
should be possible.