No need for ranges when tuning, use min of upper bound instead
The previous version of the autotuner, only guarantees optimal performance for
size-invariant parallelism, represents the optimal threshold parameter value for
each threshold and dataset as a range. After finding the threshold ranges for
each dataset, we could intersect all the ranges and pick the upper bound as our
threshold parameter value.

Since we are always just taking the upper bound, we don't need to keep track of
the ranges, just the upper bound. At the end, we can just take the minimum upper bound

Now, if such an intersection always exists, and we are always just pick the
upper bound of the intersected ranges, we don't actually need to keep track of
the ranges at all, we just need the upper bound. In essence, for a given
threshold, we are finding the best threshold parameter value for each dataset,
and then taking the minimum of all the threshold parameter values found.

The change makes intuitive sense because if the optimal threshold parameter
value for one dataset is `x`, the optimal threshold parameter value for another
dataset is `y`, and `y` < `x`, it can never hurt the performance of the first
dataset to use `y` as the threshold parameter.