We should use the best run time instead of the average run time
When benchmarking programs for the purpose of autotuning, we should optimize for
the optimal run time. Futhark is deterministic, so, all other things being
equal, a program should have the same run time each time it is run. If there's
one or more runs that have much higher run times, that's likely because of
work-sharing or something similar. Therefore, this change should result in more
consistent tuning results.