Generalize loop array returns in order to avoid unnecessary copies
The approach here is to always existentialize or generalize as much of the array
type and index function as possible. Therefore, the technique is a bit different
from the technique used to avoid copies in `if` branches, where we explicitly
try to do the least amount of generalization possible. The approach taken here
is simpler, but leads to more existentials. We therefore rely on the simplifier
to remove some of the invariant loop returns.

For this to work, we need to relax the type checker a bit. We also fixed a bug
in the imperative code generator that didn't handle ScalarSpace correctly.