Perform incremental flattening by default.
Moderate flattening is completely gone - instead, you can use
attributes to control parallelisation, if needed.

I expect this will have problems and require lots of tweaking, but it
is time we address it directly.

Compile times are harmed, of course.  I hope that is offset by the
significant improvements we've made to compile times recently (and
besides, most of the time is currently spent before kernel extraction,
during inlining, which this change does not affect).

This commit also makes minor changes to the heuristics we use to prune
the multi-versioning tree.  Nothing major, but I expect it will also
require more tuning.

Closes #697.