Set default threshold size to 2000000000 when autotuning
For incremental flattening, Futhark uses a default threshold in OpenCL GPU
programs of 32768. Some programs and benchmarks (bfast in partcular) have higher
degrees of parallelism for some combinations of threshold and dataset. To
correctly tune threshold parameters, the autotuner assumes that any unset
thresholds will always block execution of the guarded code version, but if the
default threshold is not high enough, that's not always the case. This commit
fixes that, by explicitly setting the default threshold parameter to
2000000000, which is `thresholdMax` in the tuner.

This commit should fix the autotuning problems we've been having with the bfast
benchmark.