Avoid AnyDims in entry point parameters.
This is part of a looooong yak-shaving process that aims to regularise
our handling of sizes, and in particular guarantee that there are no
anonymous sizes in function parameters coming out of
defunctionalisation.  Ideally, we will also try to retain a bit more
of the original structure, although I think we would need dependent
records for the general case.