Tiles are always partial.
This is because we forgot to take into account the kernel dimensions,
which are likely to not be divisible by the tile size.

This slightly slows down 2D-tiled programs.  A fix would be to
generate different code for workgroups at the edge of the result, but
this is beyond our current code generation facilities.

Fixes #1094.