Implement FutharkScript.
The directives are still experimental.  In particular, I wonder if we
shouldn't expose arbitrary gnuplot arguments for the :plot2d
directive.

Error handling and recovery is very basic, and apart from parse
errors, users are not particularly told *where* things go wrong.  Type
checking is also near-nonexistent.  Still, it's good enough to start
experimenting with it.

Closes #1210.