When efficiently reshaping an index funtion, keep the original base.
This makes logical sense, because the whole point of these reshape
special cases is that they only apply when the physical representation
does not matter.

It matters semantically, because otherwise the removal of a reshape
can change the base of an array, which can cause type errors (we don't
have shape-rank polymorphism).