Do not prepend all qualifiers for unreachable names.
This entire thing is bogus anyway.  The problem that we are trying to
solve is how to report the names of abstract type to users (i.e. tell
them in which modules they are defined), but we also ended up
depending on some of this for defunctorisation.  We need a better
mechanism for the former, and the latter should _only_ use the base
type name.