Handle failing sample runs correctly
If the sampling run fails, we should truly return "no information". Currently,
we return `(tMin, tMax)`, but if this is the first dataset we're tuning, and the
sample run fails, we should return `Nothing`. Then, if the collective result of
tuning a threshold is `Nothing`, we should use `thresholdMin` to select the
alternative code version. Hopefully, that won't fail.

Fixes #1336