Remove "context" part of patterns.
This is a significant simplification.  Now, if something returns N
values, its type list will have N entries, always. Probably we will be
chasing the dividends of this simplification for years, as the
context/value distinction was truly entrenched.

Entry points annotations (and some parts of the ExplicitAllocations
pass) still assume that the "context parts" (sizes) of a pattern are
first, but that is an easy property to maintain - and we might relax
it in the future.