Fix #1397.
The root cause here was that we did not remember abstract types that
were generated from subexpressions in module expressions.

I'm a little iffy about not removing the type from the abstact type
set in 'refineEnv', but it was necessary to get functor28.fut to work,
and it seems to have no ill effect.