Improve interference tracking of loop memory parameters.
We conservatively treat all memory arguments to a DoLoop to interfere
with each other, as well as anything used inside the loop.  This could
potentially be improved by looking at the interference computed by the
loop body wrt. the loop arguments, but probably very few programs
would benefit from this.