Move primitive C definitions into separate file. (#1422)
Previously we would generate these programmatically in the code
generator, but I think it's simpler to just write them once and for
all - even if that opens the door to inconsistency.  It's easier to
test and inspect this way, too.

This is also a general movement away from embedding large quasiquoted
strings in the compiler source code.