Better error message when test input has wrong type.
This only handles the (very common) case where the value is provided
explicitly.  In particular, data files are still passed directly to
the server executable.  This is because, currently, extracting the
type of the value contained within requires us to read the entire
file, which is expensive for very large input.

Closes #1441.