WIP on virtualising intragroup operations.
The idea is that the choice of whether to run intragroup should not
depend on the maximum group size (or the group size at all), but on
local memory capacity and whether it's actually fast.

Segmented intragroup reductions remain to be done, as well as
intragroup histograms at all (but I think they are pretty easy).

Also this is very lightly tested.