Support program reload cache.
I also took the liberty of cleaning up some other parts of how the
REPL works.

This is not as efficient as it could be for the REPL, since only the
type checking is cached.  However, that is good enough for a language
server, which only has the type checking part anyway.