Refactor error representation in program loading.
This is to help the work of @haoranpb on a language server.

The main difference is that a ProgramError now explicitly consists of
separate location and message, which means that a language server can
extract the former.

The API also returns potentially multiple errors, although in practice
only a few contrived cases will actually report more than a single
error.