Major overhaul of MergeGPUBodies pass.
* Feature: GPUBody statements can now be merged into even fewer kernels.
* Fix: Values within body results are now counted as dependencies.
* Reorganized code.
* Added documentation.
* Added tests.

The previous local approach of prelude/interlude/postlude groups has been replaced
with a sequence of alternating non-GPUBody/GPUBody groups. Rather than collapsing
the interlude as soon as a GPUBody cannot be merged to it the new GPUBody is now
added to a new group. This allows sequences such as:

  let a =  gpu { ... }
  let b = host { a+1 }
  let c =  gpu { b-1 }
  let d =  gpu { a*2 }
  let e = host { d+3 }
  let f =  gpu { e-1 }

to be merged into two kernels:

  let (a, d) =  gpu { ... }
  let (b, e) = host { a+1, d+3 }
  let (c, f) =  gpu { b-1, e-1 }

rather than three:

  let a      =  gpu { ... }
  let b      = host { a+1 }
  let (c, d) =  gpu { b-1, a*2 }
  let e      = host { d+3 }
  let f      =  gpu { e-1 }