Be more conservative in loop-invariant hoisting from branches.
This is to avoid hoisting out expensive stuff (e.g. replicates) from
versioning branches that might actually never be picked.  I don't like
this at all; it feels too ad-hoc.  It needs to be seriously rethought
eventually.

This optimisation was introduced for #482, and it is still necessary
for mapping of merge_sort to go well (to hoist out a size comparison).
Perhaps that can be solved by further improving the size-tracking
code.