Add futhark.scm
This is a description of the packages necessary to build Futhark with the Guix
package manager.

Unfortunately, Guix does not currently support Haskell packages with multiple
libraries in them, such as attoparsec, so we'll have to wait for
https://issues.guix.gnu.org/54729 to be merged before we can merge this.

However, with that patch we can build Futhark using `guix build -f futhark.scm`.