Switch to the 'prettyprinting' library. (#1725)
Closes #1156.

I suspect there may be cases where we don't print as nicely as we used
to, as I had to get rid of some hacks.  We can introduce them as we
notice them, as the foundations are pretty solid now.

Also, I had to make a bunch of String->Text changes all over the
place, so this looks much more invasive than it really is.