Fix #1739.
Tiling really doesn't handle irregular sizes well, so let's add a very
coarse check.

This does limit our tiling capability in a way that may hurt real
code, although I don't think I ever measured the impact of the
optimisation in this case.  It can be fixed in the future if it turns
out to be important for e.g. Gaussian elimination.