Hoist size computations even more aggressively.
Actually, what this mainly fixes is that we classify existential sizes
as sizes when simplifying the expression that produces them (mainly
Match).

This required some fixes elsewhere; mostly to prevent hoisting
thread-variant computation out of SegOp lambdas.

Ultimately this is a necessary fix to cater to the changes we made to
the allocator when we added Match, and which was hidden by a bug in
the type checker until recently.