Treat maps as short-circuit points
Current restrictions: Only treats cases where there is a single return from the
map, and a single free variable inside the map that can be used as the
destination.

Essentially, this optimization will look at code like this:

```
let ys: [n1]...[n_m] @ m_1 -> ixf_1 = {
  segmap i < n {
    ...
    let x = xs[i]
    ...
    in (x')
  }
}
```

And make ys reuse the memory of xs, if the index functions are exactly similar.