Make Futhark.Optimise.ArrayShortCircuiting.LastUse program-wide.
This is simply necessary in order for the information to be correct,
as implemented in the other pass to fix #1755.  It does not have any
effect on short-circuiting however, as we do not apply it in
constants.  (But we might one day, soon.)

The motivation for this is perhaps switching over to
Futhark.Optimise.ArrayShortCircuiting.LastUse as the sole "LastUse"
computation.