Correct number of locks for local memory case.
Previously we would create a lock for every single element, but this
is wasteful (and unnecessary) in the case of vectorised operators,
where we (at least currently) take the lock based only on the
non-vector indexes.

It might be worth changing that (as it would cause less lock
contention), and then this commit would have to be reverted.