Improve the parser somewhat.
We have long had a Great Parser Hack due to whitespace-sensitivity -
'a[x]' means something very different from 'a [x]'.  This was
implemented in a crude way, and now it is implemented in a slightly
less crude (and somewhat more robust) way.

The key idea is to treat a '[' following an identifier as a different
token from a '[' not following an identifier.

While I was rummaging around in the parser anyway, I also took the
liberty to improve our handling of projection, which is now more robust.

One difference is that it is now possible to say 'x . y' for
projection (including for module components).  I'm not certain whether
this is desirable, but it simplifies the grammar a lot to allow it.
SML allows it, too.