Eliminate FreeNode.
These were only used to capture sibling relationships between nodes
for horizontal fusion.  This is now done in a more direct (non-graph)
way.  This is probably fine, as the number of SOAC nodes in the graph
is always going to be pretty small.

The advantage of this is that fusion is quite a bit faster on large
programs, because the graph is much smaller.