Less aggressive inlining.
This commit is the beginning of reworking our inlining policy to be
less aggressive.  Currently it inlines all functions called from SOACs
that involve arrays in any way.  This is not quite enough: we also
need to inline anything that contributes to irregularity.

Given our downstream compilation restrictions, inlining isn't just
about performance - it is about whether we can compile at all.

On heston32, the impact of this is pretty promising.  Pretty much
unchanged performance, but more than 2x faster compilation.